+++ src/presenters/entity-editor.jsx
@@ -3,8 +3,6 @@ import PropTypes from 'prop-types';
 
 import {matches, reject} from 'lodash';
 
-import Uploader from './includes/uploader.jsx';
-
 export default class EntityEditor extends React.Component {
   constructor(props) {
     super(props);
@@ -47,13 +45,3 @@ EntityEditor.propTypes = {
   type: PropTypes.string.isRequired,
   children: PropTypes.func.isRequired,
 };
-
-export const EntityEditorUploader = ({api, initial, type, children}) => (
-  <EntityEditor api={api} initial={initial} type={type}>
-    {({entity, ...editFuncs}) => (
-      <Uploader>
-        {uploadFuncs => children({entity, ...editFuncs, ...uploadFuncs})}
-      </Uploader>
-    )}
-  </EntityEditor>
-);
