+++ src/presenters/includes/team-users.jsx
@@ -1,0 +1,29 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+import AddTeamUserPop from '../pop-overs/add-team-user-pop.jsx';
+import PopoverContainer from '../pop-overs/popover-container.jsx';
+import UserInfoPop from '../pop-overs/user-info-pop.jsx';
+import {UserPopoversList} from '../users-list.jsx';
+
+export const TeamUsers = ({users, currentUserIsOnTeam, removeUser}) => (
+  <UserPopoversList users={users}>
+    {(user, togglePopover) => <UserInfoPop togglePopover={togglePopover} user={user} currentUserIsOnTeam={currentUserIsOnTeam} removeUserFromTeam={() => removeUser(user.id)} />}
+  </UserPopoversList>
+);
+TeamUsers.propTypes = {
+  users: PropTypes.array.isRequired,
+  currentUserIsOnTeam: PropTypes.bool.isRequired,
+  removeUser: PropTypes.func.isRequired,
+};
+
+export const AddTeamUser = (props) => (
+  <PopoverContainer>
+    {({visible, togglePopover}) => (
+      <span className="add-user-container">
+        <button onClick={togglePopover} className="button button-small button-tertiary add-user">Add</button>
+        {visible && <AddTeamUserPop {...props} togglePopover={togglePopover}/>}
+      </span>
+    )}
+  </PopoverContainer>
+);
